+++ api/src/org/labkey/api/study/ParticipantVisit.java
@@ -30,7 +30,7 @@ import java.util.Date;
  */
 public interface ParticipantVisit
 {
-    public static final String ASSAY_RUN_MATERIAL_NAMESPACE = "AssayRunMaterial";
+    String ASSAY_RUN_MATERIAL_NAMESPACE = "AssayRunMaterial";
 
     Container getStudyContainer();
