+++ api/src/org/labkey/api/study/StudyService.java
@@ -201,8 +201,6 @@ public interface StudyService
 
     boolean runStudyImportJob(Container c, User user, ActionURL url, File studyXml, String originalFilename, BindException errors, PipeRoot pipelineRoot, ImportOptions options);
 
-    DataIteratorBuilder wrapSampleMindedTransform(User user, DataIteratorBuilder in, DataIteratorContext context, Study study, TableInfo target);
-
     ColumnInfo createAlternateIdColumn(TableInfo ti, ColumnInfo column, Container c);
 
     TableInfo getSpecimenTableUnion(QuerySchema qsDefault, Set<Container> containers, @NotNull Map<Container, SQLFragment> filterFragments, boolean dontAliasColumns, boolean useParticipantIdName);
