+++ specimen/src/org/labkey/specimen/SpecimenModule.java
@@ -35,7 +35,6 @@ import org.labkey.api.specimen.model.SpecimenDomainKind;
 import org.labkey.api.specimen.model.SpecimenEventDomainKind;
 import org.labkey.api.specimen.model.SpecimenRequestEventType;
 import org.labkey.api.specimen.model.VialDomainKind;
-import org.labkey.api.specimen.view.SpecimenReportWebPartFactory;
 import org.labkey.api.study.SpecimenService;
 import org.labkey.api.study.StudyService;
 import org.labkey.api.study.importer.SimpleStudyImporterRegistry;
@@ -45,8 +44,11 @@ import org.labkey.api.view.WebPartFactory;
 import org.labkey.specimen.action.SpecimenApiController;
 import org.labkey.specimen.importer.SpecimenSchemaImporter;
 import org.labkey.specimen.importer.SpecimenSettingsImporter;
+import org.labkey.specimen.pipeline.SampleMindedTransform;
+import org.labkey.specimen.pipeline.SampleMindedTransformTask;
 import org.labkey.specimen.security.roles.SpecimenCoordinatorRole;
 import org.labkey.specimen.security.roles.SpecimenRequesterRole;
+import org.labkey.specimen.view.SpecimenReportWebPartFactory;
 import org.labkey.specimen.view.SpecimenSearchWebPartFactory;
 import org.labkey.specimen.view.SpecimenToolsWebPartFactory;
 import org.labkey.specimen.view.SpecimenWebPartFactory;
@@ -106,10 +108,9 @@ public class SpecimenModule extends CodeOnlyModule
         ContainerManager.addContainerListener(SpecimenRequestManager.get());
 
         StudyService.get().registerStudyTabProvider(tabs->tabs.add(new SpecimensPage("Specimen Data")));
-
         SpecimenService.get().registerSpecimenImportStrategyFactory(new DefaultSpecimenImportStrategyFactory());
-
         AuditLogService.get().registerAuditType(new SpecimenCommentAuditProvider());
+        SpecimenService.get().registerSpecimenTransform(new SampleMindedTransform());
 
         SimpleStudyWriterRegistry.registerSimpleStudyWriterProvider(() -> List.of(
             new SpecimenSettingsWriter(),
@@ -134,7 +135,8 @@ public class SpecimenModule extends CodeOnlyModule
     public Set<Class> getUnitTests()
     {
         return Set.of(
-            SpecimenWriter.TestCase.class
+            SpecimenWriter.TestCase.class,
+            SampleMindedTransformTask.TestCase.class
         );
     }
 }
